# api/openapi.yaml
openapi: 3.1.0
info:
  title: Seva API
  description: |
    API specification for the Seva platform.
    This is the single source of truth that generates Go server stubs,
    SvelteKit TypeScript client, and Flutter Dart client.
  version: 1.0.0
  contact:
    name: Platform Engineering
    email: engineering@seva.io
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.staging.seva.io
    description: Staging
  - url: https://api.seva.io
    description: Production
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/users/me:
    get:
      operationId: getCurrentUser
      summary: Get current user profile
      description: Returns the profile of the currently authenticated user.
      tags: [Users]
      responses:
        "200":
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
              example:
                id: "usr_01HQXK5V8R3N0E7YGZM2PFWT4A"
                type: customer
                phone: "+919876543210"
                email: priya@example.com
                name: Priya Sharma
                jurisdiction_id: "IN-KA"
                preferred_language: en
                device_type: android
                created_at: "2026-01-15T08:30:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      operationId: updateCurrentUser
      summary: Update current user profile
      description: Updates the profile of the currently authenticated user.
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
            example:
              name: Priya S.
              email: priya.s@example.com
              preferred_language: kn
      responses:
        "200":
          description: User profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/users/{id}:
    get:
      operationId: getUserById
      summary: Get user by ID
      description: Returns a user profile by ID. Restricted to admin users.
      tags: [Users]
      parameters:
        - $ref: "#/components/parameters/UserIdPath"
      responses:
        "200":
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ---------------------------------------------------------------------------
  # Providers
  # ---------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from the /auth/verify-otp or /auth/refresh endpoints.

  # ---------------------------------------------------------------------------
  # Reusable Parameters
  # ---------------------------------------------------------------------------
  parameters:
    UserIdPath:
      name: id
      in: path
      required: true
      description: User ID
      schema:
        type: string
      example: "usr_01HQXK5V8R3N0E7YGZM2PFWT4A"
...
  responses:
    BadRequest:
      description: Bad request — invalid input or missing required fields
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: BAD_REQUEST
            message: Invalid request parameters
            details:
              - field: phone
                message: Phone number must include country code
    Unauthorized:
      description: Unauthorized — missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: UNAUTHORIZED
            message: Authentication required
            details: null
    Forbidden:
      description: Forbidden — insufficient permissions for the requested action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: FORBIDDEN
            message: You do not have permission to perform this action
            details: null
    NotFound:
      description: Not found — the requested resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: NOT_FOUND
            message: Resource not found
            details: null
    UnprocessableEntity:
      description: Unprocessable entity — the request is well-formed but semantically invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: UNPROCESSABLE_ENTITY
            message: Validation failed
            details:
              - field: rating
                message: Rating must be between 1 and 5
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: INTERNAL_SERVER_ERROR
            message: An unexpected error occurred
            details: null

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    UserType:
      type: string
      enum: [customer, provider, admin]
      description: The role type of the user
    DeviceType:
      type: string
      enum: [android, ios, web]
      description: Device type the user registered from
    User:
      type: object
      required: [id, type, phone, name, jurisdiction_id, created_at]
      properties:
        id:
          type: string
          description: Unique user identifier (ULID)
          examples:
            - "usr_01HQXK5V8R3N0E7YGZM2PFWT4A"
        type:
          $ref: "#/components/schemas/UserType"
        phone:
          type: string
          description: Phone number (E.164 format)
          examples:
            - "+919876543210"
        email:
          type: string
          format: email
          description: Email address (optional)
          examples:
            - "priya@example.com"
        name:
          type: string
          description: Full name
          minLength: 1
          maxLength: 200
          examples:
            - "Priya Sharma"
        jurisdiction_id:
          type: string
          description: Jurisdiction identifier (ISO 3166-2)
          examples:
            - "IN-KA"
        preferred_language:
          type: string
          description: Preferred language (ISO 639-1)
          default: en
          examples:
            - "en"
        device_type:
          $ref: "#/components/schemas/DeviceType"
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
    UserUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        email:
          type: string
          format: email
        preferred_language:
          type: string
          description: ISO 639-1 language code
        device_type:
          $ref: "#/components/schemas/DeviceType"

    # --- Providers ---
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Machine-readable error code
          examples:
            - "BAD_REQUEST"
        message:
          type: string
          description: Human-readable error message
          examples:
            - "Invalid request parameters"
        details:
          description: Additional error details (validation errors, etc.)
          oneOf:
            - type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
            - type: "null"

    # --- Notifications ---
...